pyproject.toml
[tool.poetry]
name = "crescent-ext-prefixes"
version = "0.1.0"
description = "A library that lets you write prefix commands with crescent."
authors = ["Lunarmagpie <Bambolambo0@gmail.com>"]
license = "MPL-2.0"
packages = [
    { include = "crescent/ext/prefixes" }
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
hikari-crescent = "^0.3.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "^0.961"

[tool.mypy]
strict = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
